requirements.txt

---

## 🧰 **2. requirements.txt**

```text
Django==5.0.2
djangorestframework==3.15.2
django-crispy-forms==2.2
django-tailwind==3.8.0
Pillow==10.0.0
django-filter==24.2
drf-yasg==1.21.7
python-dotenv==1.0.1
